requirements.txt
cffi==1.15.0
chardet==4.0.0
cryptography==35.0.0
pdfminer.six==20211012
pycparser==2.20
PyYAML==6.0
